# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SaaS Factory API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ProjectStatus = typeof ProjectStatus[keyof typeof ProjectStatus];


export const ProjectStatus = {
  draft: 'draft',
  generating: 'generating',
  ready: 'ready',
  deployed: 'deployed',
  error: 'error',
} as const;

export interface Project {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  slug: string;
  status: ProjectStatus;
  template: string;
  creditsUsed: number;
  /** @nullable */
  deployedUrl?: string | null;
  /** @nullable */
  previewUrl?: string | null;
  /** @nullable */
  generatedCode?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface ProjectInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  template: string;
}

export interface ProjectUpdate {
  /** @minLength 1 */
  name?: string;
  description?: string;
  status?: string;
  generatedCode?: string;
}

export interface ProjectFile {
  id: number;
  projectId: number;
  filePath: string;
  content: string;
  language: string;
  createdAt: string;
}

export type DeploymentPlatform = typeof DeploymentPlatform[keyof typeof DeploymentPlatform];


export const DeploymentPlatform = {
  vercel: 'vercel',
  netlify: 'netlify',
  replit: 'replit',
} as const;

export type DeploymentStatus = typeof DeploymentStatus[keyof typeof DeploymentStatus];


export const DeploymentStatus = {
  pending: 'pending',
  building: 'building',
  live: 'live',
  failed: 'failed',
} as const;

export interface Deployment {
  id: number;
  projectId: number;
  platform: DeploymentPlatform;
  status: DeploymentStatus;
  /** @nullable */
  deployedUrl?: string | null;
  /** @nullable */
  buildLogs?: string | null;
  createdAt: string;
}

export type DeploymentInputPlatform = typeof DeploymentInputPlatform[keyof typeof DeploymentInputPlatform];


export const DeploymentInputPlatform = {
  vercel: 'vercel',
  netlify: 'netlify',
  replit: 'replit',
} as const;

export interface DeploymentInput {
  platform: DeploymentInputPlatform;
}

export interface Template {
  id: string;
  name: string;
  description: string;
  category: string;
  icon: string;
  popular?: boolean;
}

export interface GenerationInput {
  /** @minLength 1 */
  prompt: string;
  model?: string;
}

export interface GenerationResult {
  success: boolean;
  message: string;
  creditsUsed: number;
  /** @nullable */
  generatedCode?: string | null;
}

export type DashboardStatsProjectsByStatus = {[key: string]: number};

export interface DashboardStats {
  totalProjects: number;
  deployedProjects: number;
  creditsRemaining: number;
  creditsUsed: number;
  activeGenerations: number;
  projectsByStatus?: DashboardStatsProjectsByStatus;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  created: 'created',
  generated: 'generated',
  deployed: 'deployed',
  error: 'error',
  updated: 'updated',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  projectName: string;
  /** @nullable */
  projectId?: number | null;
  createdAt: string;
}