# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SaaS Factory API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: projects
    description: Project management
  - name: templates
    description: SaaS templates
  - name: deployments
    description: Deployment management
  - name: generation
    description: AI generation
  - name: dashboard
    description: Dashboard statistics

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List all projects
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"

  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get a project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: Not found
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /projects/{id}/generate:
    post:
      operationId: generateProject
      tags: [generation]
      summary: Trigger AI generation for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GenerationInput"
      responses:
        "200":
          description: Generation started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationResult"

  /projects/{id}/deploy:
    post:
      operationId: deployProject
      tags: [deployments]
      summary: Deploy a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentInput"
      responses:
        "201":
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"

  /projects/{id}/deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List deployments for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"

  /projects/{id}/files:
    get:
      operationId: listProjectFiles
      tags: [projects]
      summary: List files for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectFile"

  /templates:
    get:
      operationId: listTemplates
      tags: [templates]
      summary: List all available templates
      responses:
        "200":
          description: List of templates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Template"

  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary statistics
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"

  /dashboard/activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      responses:
        "200":
          description: Recent activity items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Project:
      type: object
      required: [id, name, slug, status, template, creditsUsed, createdAt, updatedAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        slug:
          type: string
        status:
          type: string
          enum: [draft, generating, ready, deployed, error]
        template:
          type: string
        creditsUsed:
          type: integer
        deployedUrl:
          type: ["string", "null"]
        previewUrl:
          type: ["string", "null"]
        generatedCode:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string

    ProjectInput:
      type: object
      required: [name, template]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        template:
          type: string

    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
        generatedCode:
          type: string

    ProjectFile:
      type: object
      required: [id, projectId, filePath, content, language, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        filePath:
          type: string
        content:
          type: string
        language:
          type: string
        createdAt:
          type: string

    Deployment:
      type: object
      required: [id, projectId, platform, status, createdAt]
      properties:
        id:
          type: integer
        projectId:
          type: integer
        platform:
          type: string
          enum: [vercel, netlify, replit]
        status:
          type: string
          enum: [pending, building, live, failed]
        deployedUrl:
          type: ["string", "null"]
        buildLogs:
          type: ["string", "null"]
        createdAt:
          type: string

    DeploymentInput:
      type: object
      required: [platform]
      properties:
        platform:
          type: string
          enum: [vercel, netlify, replit]

    Template:
      type: object
      required: [id, name, description, category, icon]
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
        category:
          type: string
        icon:
          type: string
        popular:
          type: boolean

    GenerationInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
        model:
          type: string

    GenerationResult:
      type: object
      required: [success, message, creditsUsed]
      properties:
        success:
          type: boolean
        message:
          type: string
        creditsUsed:
          type: integer
        generatedCode:
          type: ["string", "null"]

    DashboardStats:
      type: object
      required: [totalProjects, deployedProjects, creditsRemaining, creditsUsed, activeGenerations]
      properties:
        totalProjects:
          type: integer
        deployedProjects:
          type: integer
        creditsRemaining:
          type: integer
        creditsUsed:
          type: integer
        activeGenerations:
          type: integer
        projectsByStatus:
          type: object
          additionalProperties:
            type: integer

    ActivityItem:
      type: object
      required: [id, type, message, projectName, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [created, generated, deployed, error, updated]
        message:
          type: string
        projectName:
          type: string
        projectId:
          type: ["integer", "null"]
        createdAt:
          type: string