# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SaaS Factory API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsProjectsByStatus } from './dashboardStatsProjectsByStatus';

export interface DashboardStats {
  totalProjects: number;
  deployedProjects: number;
  creditsRemaining: number;
  creditsUsed: number;
  activeGenerations: number;
  projectsByStatus?: DashboardStatsProjectsByStatus;
}