# lib/api-zod/src/generated/types/deployment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SaaS Factory API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeploymentPlatform } from './deploymentPlatform';
import type { DeploymentStatus } from './deploymentStatus';

export interface Deployment {
  id: number;
  projectId: number;
  platform: DeploymentPlatform;
  status: DeploymentStatus;
  /** @nullable */
  deployedUrl?: string | null;
  /** @nullable */
  buildLogs?: string | null;
  createdAt: string;
}