# lib/api-zod/src/generated/types/deploymentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SaaS Factory API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DeploymentInputPlatform } from './deploymentInputPlatform';

export interface DeploymentInput {
  platform: DeploymentInputPlatform;
}