# lib/api-zod/src/generated/types/generationInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SaaS Factory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerationInput {
  /** @minLength 1 */
  prompt: string;
  model?: string;
}