# lib/api-zod/src/generated/types/generationResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SaaS Factory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface GenerationResult {
  success: boolean;
  message: string;
  creditsUsed: number;
  /** @nullable */
  generatedCode?: string | null;
}