# lib/api-zod/src/generated/types/projectFile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SaaS Factory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ProjectFile {
  id: number;
  projectId: number;
  filePath: string;
  content: string;
  language: string;
  createdAt: string;
}