# lib/api-zod/src/generated/types/template.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SaaS Factory API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Template {
  id: string;
  name: string;
  description: string;
  category: string;
  icon: string;
  popular?: boolean;
}